Separate the components into containers and components
